#817: fix interdependency between mobileKeyword and journey on auto-formatting
