+++ lib/metadataTypes/MobileKeyword.js
@@ -4,7 +4,6 @@ const TYPE = require('../../types/mcdev.d');
 const MetadataType = require('./MetadataType');
 const Util = require('../util/util');
 const File = require('../util/file');
-const beautifier = require('beauty-amp-core');
 const cache = require('../util/cache');
 
 /**
@@ -142,26 +141,7 @@ class MobileKeyword extends MetadataType {
      * @returns {{fileExt:string,code:string}} returns found extension and file content
      */
     static prepExtractedCode(metadataScript) {
-        // immutable at the moment:
-        const ampscript = {
-            capitalizeAndOrNot: true,
-            capitalizeIfFor: true,
-            capitalizeSet: true,
-            capitalizeVar: true,
-            maxParametersPerLine: 4,
-        };
-        // immutable at the moment:
-        const editor = {
-            insertSpaces: true,
-            tabSize: 4,
-        };
-        // logs trough console only for the moment.
-        const logs = {
-            loggerOn: false, // <= disable logging
-        };
-
-        beautifier.setup(ampscript, editor, logs);
-        const code = beautifier.beautify(metadataScript);
+        const code = metadataScript;
         const fileExt = 'amp';
 
         return { fileExt, code };
