+++ lib/util/file.js
@@ -6,6 +6,7 @@ const fs = require('fs-extra');
 const packageJson = require('../../package.json');
 const path = require('node:path');
 const prettier = require('prettier');
+const beautyAmp = require('beauty-amp-core');
 const Util = require('./util');
 const updateNotifier = require('update-notifier');
 
@@ -168,11 +169,46 @@ const File = {
      * @returns {Promise.<boolean>} Promise
      */
     writePrettyToFile: async function (directory, filename, filetype, content, templateVariables) {
-        let formatted = await this._beautify_prettier(directory, filename, filetype, content);
+        let formatted =
+            filetype === 'amp'
+                ? this._beautify_beautyAmp(content)
+                : await this._beautify_prettier(directory, filename, filetype, content);
         if (templateVariables) {
             formatted = Util.replaceByObject(formatted, templateVariables);
         }
         return this.writeToFile(directory, filename, filetype, formatted);
+    },
+    /**
+     * helper for {@link File.writePrettyToFile}, applying beautyAmp onto given stringified content
+     *
+     * @param {string} content filecontent
+     * @returns {string} original string on error; formatted string on success
+     */
+    _beautify_beautyAmp: function (content) {
+        // immutable at the moment:
+        const ampscript = {
+            capitalizeAndOrNot: true,
+            capitalizeIfFor: true,
+            capitalizeSet: true,
+            capitalizeVar: true,
+            maxParametersPerLine: 4,
+        };
+        // immutable at the moment:
+        const editor = {
+            insertSpaces: true,
+            tabSize: 4,
+        };
+        // logs trough console only for the moment.
+        const logs = {
+            loggerOn: false, // <= disable logging
+        };
+        try {
+            beautyAmp.setup(ampscript, editor, logs);
+            return beautyAmp.beautify(content);
+        } catch (ex) {
+            Util.logger.debug('File._beautify_beautyAmp:: error | ' + ex.message);
+            return content;
+        }
     },
     /**
      * helper for {@link File.writePrettyToFile}, applying prettier onto given stringified content
@@ -185,16 +221,16 @@ const File = {
      * @returns {string} original string on error; formatted string on success
      */
     _beautify_prettier: function (directory, filename, filetype, content) {
-        if (!FileFs.prettierConfig) {
-            // either no prettier config in project directory or initPrettier was not run before this
-            return content;
-        } else if (content.includes('%%[') || content.includes('%%=')) {
-            // in case we find AMPScript we need to abort beautifying as prettier
-            // will throw an error falsely assuming bad syntax
-            return content;
-        }
         let formatted = '';
         try {
+            if (!FileFs.prettierConfig) {
+                // either no prettier config in project directory or initPrettier was not run before this
+                return content;
+            } else if (content.includes('%%[') || content.includes('%%=')) {
+                // in case we find AMPScript we need to abort beautifying as prettier
+                // will throw an error falsely assuming bad syntax
+                return this._beautify_beautyAmp(content);
+            }
             // load the right prettier config relative to our file
             switch (filetype) {
                 case 'htm':
